LOCAL modules: enable rtw89 and ath1[012]k, add iwn and iwm to arm64 too
It seems there is a chance iwn and iwm just work on arm64 so enable
it there.  It certainly helps me to do more testing.  I've had partial
success with them and it is unclear if the problem remaining is my
setup.  If committed I would likely not enable iwm(4) on arm64 anymore.

rtw89 does compile and scan but not yet pass packet but in order for
people to try and test and possibly debug a bit, enable it here.

Enable athk_common, ath10k, ath11k, ath12k modules.
So far the drivers do not properly attach (or are missing code for
ath11k and ath12k) but make sure we keep compiling them at least until
time is found to debug and make them work.

Sponsored by:	The FreeBSD Foundation
MFC after:	3 days