iwlwifi: disable request_module() for iwlmvm
In Linux iwlwifi seems to be multiple modules, a base iwlwifi and at
least iwlmvm (which we use) and iwldvm (which matches our iwn(4)).

There are cases with multiple iwlwifi cards, for me on arm64, when we
try to load, and the multiple times probe and attach the driver.
In these circumstances it seems to happen that request_module() which
calls kern_kldload() triggered from kldload seems to be able to hang
somehow.

Given the iwlmvm parts are part of if_iwlwifi.ko on FreeBSD there is
no need to try to load the extra module.

With this I manage to probe up to 7 iwlwifi cards currently in a single
machine (I have not tried more).

Sponsord by:	The FreeBSD Foundation
MFC after:	3 days