fwget: update wireless IDs for rtw88/89 and mt76 and add iwlwifi
Update and add (new) PCI IDs for Realtek rtw88/89, Mediatek 7996/7925
and add Intel iwlwifi IDs.

For the latter we need a great level of detail so also pass the full
pciconf -l line to into the pci_* files as "$2" to have access to these.

Sponsored by:	The FreeBSD Foundation
MFC after:	3 days